ov065: promote daObjCtMecha04_c into a single C++ TU
