+++ attribution.json
@@ -1768,7 +1768,6 @@
     "src/_ZN14dScMgMemory2_c11InitPlayersEv.c": "github-actions[bot]",
     "src/_ZN14dScMgMemory2_c8CardWaitEi.c": "github-actions[bot]",
     "src/_ZN14dScMgMemory2_c8CardIdleEi.c": "github-actions[bot]",
-    "src/_ZN14dScMgMemory2_c14RoundShowCardsEv.c": "github-actions[bot]",
     "src/_ZN14dScMgMemory2_c10StateJudgeEv.c": "github-actions[bot]",
     "src/_ZN14dScMgMemory2_c10StateSetupEv.c": "github-actions[bot]",
     "src/_ZN14dScMgMemory2_c10PlayerWaitEi.c#func_ov006_020f5740": "github-actions[bot]",
@@ -1958,6 +1957,15 @@
     "src/game/actors/d_a_obj_rc_hane.cpp#func_ov036_02111618": "github-actions[bot]",
     "src/game/actors/d_a_obj_rc_hane.cpp#_ZN14daObjRc_Hane_c6RenderEv": "andrewboudreau",
     "src/game/actors/d_a_obj_rc_hane.cpp#_ZN14daObjRc_Hane_c8BehaviorEv": "github-actions[bot]",
-    "src/game/actors/d_a_obj_rc_hane.cpp#_ZN14daObjRc_Hane_c13InitResourcesEv": "github-actions[bot]"
+    "src/game/actors/d_a_obj_rc_hane.cpp#_ZN14daObjRc_Hane_c13InitResourcesEv": "github-actions[bot]",
+    "src/actors/daObjCtMecha04_c.cpp#_ZN16daObjCtMecha04_cD1Ev": "github-actions[bot]",
+    "src/actors/daObjCtMecha04_c.cpp#_ZN16daObjCtMecha04_cD0Ev": "github-actions[bot]",
+    "src/actors/daObjCtMecha04_c.cpp#_ZN16daObjCtMecha04_c12UpdateShadowEv": "tangosdev",
+    "src/actors/daObjCtMecha04_c.cpp#_ZN16daObjCtMecha04_c16CleanupResourcesEv": "github-actions[bot]",
+    "src/actors/daObjCtMecha04_c.cpp#_ZN16daObjCtMecha04_c6RenderEv": "github-actions[bot]",
+    "src/actors/daObjCtMecha04_c.cpp#_ZN16daObjCtMecha04_c8BehaviorEv": "andrewboudreau",
+    "src/actors/daObjCtMecha04_c.cpp#_ZN16daObjCtMecha04_c13InitResourcesEv": "andrewboudreau",
+    "src/actors/daObjCtMecha04_c.cpp#_ZN16daObjCtMecha04_c15MoveActorOnBeltER8dActor_c": "noreply",
+    "src/actors/daObjCtMecha04_c.cpp#_ZN16daObjCtMecha04_c17AfterClsnCallbackEP4dBgWP8dActor_cS3_": "tangosdev"
   }
 }
