+++ config/converted-baseline.json
@@ -890,12 +890,12 @@
     "src/_ZN16daObjCtMecha03_c6RenderEv.cpp",
     "src/_ZN16daObjCtMecha03_cD0Ev.cpp",
     "src/_ZN16daObjCtMecha03_cD1Ev.cpp",
-    "src/_ZN16daObjCtMecha04_c15MoveActorOnBeltER8dActor_c.cpp",
-    "src/_ZN16daObjCtMecha04_c16CleanupResourcesEv.cpp",
-    "src/_ZN16daObjCtMecha04_c17AfterClsnCallbackEP4dBgWP8dActor_cS3_.cpp",
-    "src/_ZN16daObjCtMecha04_c6RenderEv.cpp",
-    "src/_ZN16daObjCtMecha04_cD0Ev.cpp",
-    "src/_ZN16daObjCtMecha04_cD1Ev.cpp",
+    "src/actors/daObjCtMecha04_c.cpp#_ZN16daObjCtMecha04_c15MoveActorOnBeltER8dActor_c",
+    "src/actors/daObjCtMecha04_c.cpp#_ZN16daObjCtMecha04_c16CleanupResourcesEv",
+    "src/actors/daObjCtMecha04_c.cpp#_ZN16daObjCtMecha04_c17AfterClsnCallbackEP4dBgWP8dActor_cS3_",
+    "src/actors/daObjCtMecha04_c.cpp#_ZN16daObjCtMecha04_c6RenderEv",
+    "src/actors/daObjCtMecha04_c.cpp#_ZN16daObjCtMecha04_cD0Ev",
+    "src/actors/daObjCtMecha04_c.cpp#_ZN16daObjCtMecha04_cD1Ev",
     "src/_ZN16daObjFallBlock_c15OnHitByMegaCharER6Player.cpp",
     "src/_ZN16daObjFallBlock_c6RenderEv.cpp",
     "src/_ZN16daObjFallBlock_cD0Ev.cpp",
