--- src/_ZN16daObjCtMecha04_cD0Ev.cpp
@@ -1,9 +1,0 @@
-//cpp
-// @symbol _ZN16daObjCtMecha04_cD0Ev
-/* D0 is the deleting wrapper emitted beside the same real destructor. The
- * inherited inline operator delete returns the actor to data_020a0eac. */
-#include "daObjCtMecha04_c.h"
-
-daObjCtMecha04_c::~daObjCtMecha04_c()
-{
-}
