--- src/_ZN16daObjCtMecha04_cD1Ev.cpp
@@ -1,10 +1,0 @@
-//cpp
-// @symbol _ZN16daObjCtMecha04_cD1Ev
-/* The compiler emits the complete destructor from the real base and member
- * declarations: ShadowModel, TextureTransformer, then the inlined
- * dBgActor_c chain. */
-#include "daObjCtMecha04_c.h"
-
-daObjCtMecha04_c::~daObjCtMecha04_c()
-{
-}
