+++ symbols/actor_renames.tsv
@@ -3171,8 +3171,8 @@ ov002	0x020b42e4	func_ov002_020b42e4	_ZN10BrickBlock15SpawnSilverStarEv	PMF acti
 ov002	0x020b4714	func_ov002_020b4714	_ZN16daObjKinokoTag_c17SpawnMegaMushroomEv	inferred helper name; class name from ROM RTTI and this ownership proven by exclusive TU callers and field ABI
 ov002	0x020b47ec	func_ov002_020b47ec	_ZN16daObjKinokoTag_c20TrySpawnMegaMushroomEv	inferred helper name; class name from ROM RTTI and this ownership proven by exclusive Behavior caller and collider field ABI
 ov065	0x0211a550	func_ov065_0211a550	_ZN16daObjCtMecha04_c12UpdateShadowEv	inferred helper name; class name from ROM RTTI and this ownership proven by exclusive Behavior caller, owned shadow fields, and dActor_c shadow call
-ov065	0x0211aa38	func_ov065_0211aa38	_ZN16daObjCtMecha04_c15MoveActorOnBeltER8dActor_c	inferred helper name and dActor_c reference parameter; callback forwarding and conveyor/actor field ABI prove ownership and call shape
-ov065	0x0211aacc	func_ov065_0211aacc	_ZN16daObjCtMecha04_c17AfterClsnCallbackEP4dBgWP8dActor_cS3_	inferred helper name and callback types; installed as the collider callback and observed forwarding owner and colliding actor
+ov065	0x0211aa38	func_ov065_0211aa38	_ZN16daObjCtMecha04_c15MoveActorOnBeltER8dActor_c	English helper name, member/free form, target type, and dActor_c& spelling are inferred; callback forwarding and conveyor/actor field ABI prove ownership and call shape
+ov065	0x0211aacc	func_ov065_0211aacc	_ZN16daObjCtMecha04_c17AfterClsnCallbackEP4dBgWP8dActor_cS3_	English helper name, static/member/free form, and pointer/type spellings are inferred; installation and forwarding registers prove the collider callback role
 ov072	0x02120c00	func_ov072_02120c00	daBgSnwmn_c_classInit	reconstructed classInit; ROM RTTI+factory shape+unique class factory+later EAD lineage; exact SM64DS spelling not preserved
 ov072	0x02122954	data_ov072_02122954	g_profile_BIG_SNOWMAN	reconstructed profile global; literal ROM ID BIG_SNOWMAN+registry role+later EAD lineage; exact SM64DS spelling not preserved
 ov072	0x02122978	data_ov072_02122978	_ZTV11daBgSnwmn_c	ROM RTTI class identity; address-point symbol already configured at the same address
